nodeos_alert/requirements.txt
requests==2.19.1
python-dateutil==2.6.0